fix comms console removing red alert button on red alert
