+++ Content.Client/Communications/UI/CommunicationsConsoleMenu.xaml
@@ -1,14 +1,3 @@
-<!--
-SPDX-FileCopyrightText: 2023 DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>
-SPDX-FileCopyrightText: 2023 Morb <14136326+Morb0@users.noreply.github.com>
-SPDX-FileCopyrightText: 2024 Red Mushie <82113471+redmushie@users.noreply.github.com>
-SPDX-FileCopyrightText: 2024 VideoKompany <135313844+VlaDOS1408@users.noreply.github.com>
-SPDX-FileCopyrightText: 2024 avery <51971268+graevy@users.noreply.github.com>
-SPDX-FileCopyrightText: 2025 Aiden <28298836+Aidenkrz@users.noreply.github.com>
-
-SPDX-License-Identifier: AGPL-3.0-or-later
--->
-
 <controls:FancyWindow xmlns="https://spacestation14.io"
     xmlns:controls="clr-namespace:Content.Client.UserInterface.Controls"
     Title="{Loc 'comms-console-menu-title'}"
